backend/auth/views/UpdateUserPasswordView.py
from django.contrib.auth.models import User
from rest_framework.views import APIView
from rest_framework.response import Response
from rest_framework.permissions import IsAuthenticated

class UpdateUserPasswordView(APIView):
    permission_classes = [IsAuthenticated]

    def post(self, request):
        user = request.user
        new_password = request.data.get('new_password')

        # Update the password
        user.set_password(new_password)
        user.save()

        return Response({"message": "Password updated successfully."})